Refactor all endpoints into alert_mailer
