+++ app/controllers/api/v1/alert_mailer_controller.rb
@@ -1,11 +1,9 @@
-class Api::V1::AlertsController < ApplicationController
-  def index
+class Api::V1::AlertMailerController < ApplicationController
+  def create
     @alerts = NwsFacade.escambia_alerts
     if @alerts.blank?
       render json: { data: 'No Current Alerts' }
     else
-      render json: { data: 'Current Alerts in Your Area' }
-
       # Create the user from params
       @user = User.find_or_create_by(user_params)
 
@@ -13,7 +11,7 @@ class Api::V1::AlertsController < ApplicationController
       alerts = NwsFacade.escambia_alerts
       tweets = [TwitterFacade.fl511_alerts, TwitterFacade.bereadyescambia_alerts]
       UserNotifierMailer.send_alerts(@user, alerts, tweets).deliver_now
-      redirect_to(api_v1_users_url(@user), notice: 'User created')
+      render json: { data: 'Current Alerts in Your Area' }
     end
   end
