+++ config/routes.rb
@@ -1,8 +1,7 @@
 Rails.application.routes.draw do
   namespace :api do
     namespace :v1 do
-      resources :users, only: [:index, :show, :create]
-      get 'EscambiaCountyFLAlerts', to: 'nws_alerts#index'
+      post '/alert_mailer', to: 'alert_mailer#create'
     end
   end
 end
