+++ spec/requests/api/v1/alert_mailer/create_spec.rb
@@ -1,0 +1,69 @@
+require 'rails_helper'
+
+RSpec.describe 'Create Alert Mailer API' do
+  context 'When there are alert records' do
+    describe 'can POST a new alert mailer' do
+      before :each do
+        json_response = File.read('spec/fixtures/DO_NOT_DELETE/services/alerts_with_moderate_filter.json')
+        stub_request(:get, 'https://api.weather.gov/alerts?zone=FLZ202&status=actual&severity=Severe,Extreme').to_return(
+          status: 200, body: json_response
+        )
+
+        post api_v1_alert_mailer_path(:user => {name: 'mufasa', email: 'mufasa@whatever.com'})
+      end
+
+      let!(:alerts) { JSON.parse(response.body, symbolize_names: true) }
+
+      it 'returns http status code 200' do
+        expect(response).to be_successful
+        expect(response).to have_http_status(200)
+      end
+
+      it 'returns alerts' do
+        expect(alerts).to be_a Hash
+        expect(alerts).to have_key(:data)
+        expect(alerts[:data]).to eq('Current Alerts in Your Area')
+      end
+
+      context 'with valid parameters' do
+        let(:valid_attributes) do
+          {
+            name: Faker::Name.first_name,
+            email: Faker::Internet.email
+          }
+        end
+        let(:valid_headers) do
+          {}
+        end
+
+        it 'creates a new User' do
+          expect do
+            post api_v1_alert_mailer_url,
+                params: { user: valid_attributes }, headers: valid_headers, as: :json
+          end.to change(User, :count).by(1)
+        end
+      end
+    end
+  end
+
+  context 'When there are NO alert records' do
+    before :each do
+      json_response = File.read('spec/fixtures/DO_NOT_DELETE/services/no_current_alerts.json')
+      stub_request(:get, 'https://api.weather.gov/alerts?zone=FLZ202&status=actual&severity=Severe,Extreme').to_return(
+        status: 200, body: json_response
+      )
+      post api_v1_alert_mailer_path
+    end
+
+    let!(:alerts) { JSON.parse(response.body, symbolize_names: true) }
+
+    it 'returns http status 200' do
+      expect(response).to have_http_status(200)
+    end
+
+    it 'returns no current alerts' do
+      expect(alerts).to have_key(:data)
+      expect(alerts[:data]).to eq('No Current Alerts')
+    end
+  end
+end
