--- spec/requests/api/v1/routing/users_routing_spec.rb
@@ -1,17 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe Api::V1::UsersController, type: :routing do
-  describe 'routing' do
-    it 'routes to #index' do
-      expect(get: '/api/v1/users').to route_to('api/v1/users#index')
-    end
-
-    it 'routes to #show' do
-      expect(get: '/api/v1/users/1').to route_to('api/v1/users#show', id: '1')
-    end
-
-    it 'routes to #create' do
-      expect(post: '/api/v1/users').to route_to('api/v1/users#create')
-    end
-  end
-end
